automate/golocal_sharepoint_pricingengine/requirements.txt
pandas
pytz
datetime
office365-python-client
email
smtplib
python-dotenv
google-cloud-bigquery